requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Werkzeug==2.3.7
PyMongo==4.4.0
requests
beautifulsoup4
python-dotenv
dnspython # required for atlas
gunicorn # for production deployment
